Update group_stories.mustache
The width is currently hardcoded to 3. This works fine for the examples in tests, but for use cases with more than 3 values in the examples table the format is not correct. Using the header size ensures the formatting is correct regardless of the number of values in the examples table.